空のディレクトリ生成を追加

+++ init.py
@@ -17,12 +17,22 @@ urllib.request.urlretrieve(url, save_name)
 with zipfile.ZipFile("./CodeIgniter.zip") as existing_zip:
     existing_zip.extractall("./")
 
+# mkdir
+if not os.path.exists("./html"):
+    os.makedirs("./html")
+if not os.path.exists("./mysql/data"):
+    os.makedirs("./mysql/data")
+if not os.path.exists("./phpmyadmin/sessions"):
+    os.makedirs("./phpmyadmin/sessions")
+
 # copy
 source_dir = "./CodeIgniter-" + CODE_IGNITER_VER
 target_dir = "./html"
 shutil.copy(source_dir + "/index.php", target_dir + "/index.php")
-shutil.copytree(source_dir + "/application", target_dir + "/application")
-shutil.copytree(source_dir + "/system", target_dir + "/system")
+if not os.path.exists(target_dir + "/application"):
+    shutil.copytree(source_dir + "/application", target_dir + "/application")
+if not os.path.exists(target_dir + "/system"):
+    shutil.copytree(source_dir + "/system", target_dir + "/system")
 
 # delete
 os.remove("./CodeIgniter.zip")
